Added check for directory existence
When upper and lower mounts do not exist, the tests fail to execute.
This patch allows tests to run successfully. There is also a change in
name of mount type from 'overlayfs' to 'overlay'. With these changes
tests pass successfully. The --um doesn't work as it requires specific
mount binary to be present in a specific path.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>